+++ packages/cloud_firestore/cloud_firestore_web/lib/src/interop/firestore_interop.dart
@@ -6,7 +6,7 @@
 // ignore_for_file: public_member_api_docs
 
 @JS('firebase_firestore')
-library firebase_interop.firestore;
+library;
 
 import 'dart:js_interop';