+++ packages/cloud_functions/cloud_functions/lib/src/https_callable.dart
@@ -3,7 +3,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-part of cloud_functions;
+part of '../cloud_functions.dart';
 
 /// A reference to a particular Callable HTTPS trigger in Cloud Functions.
 ///