+++ packages/firebase_auth/firebase_auth/lib/src/user.dart
@@ -3,7 +3,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-part of firebase_auth;
+part of '../firebase_auth.dart';
 
 /// A user account.
 class User {
@@ -673,7 +673,9 @@ class User {
   }
 
   MultiFactor get multiFactor {
-    if (!kIsWeb && (Platform.isMacOS || Platform.isWindows)) {
+    if (!kIsWeb &&
+        (defaultTargetPlatform == TargetPlatform.macOS ||
+            defaultTargetPlatform == TargetPlatform.windows)) {
       throw UnimplementedError(
         'MultiFactor Authentication is only supported on web, Android and iOS.',
       );