+++ packages/firebase_data_connect/firebase_data_connect/lib/firebase_data_connect.dart
@@ -12,9 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-library firebase_data_connect;
-
-export 'src/firebase_data_connect.dart';
+export 'src/any_value.dart' show AnyValue, defaultSerializer;
 export 'src/common/common_library.dart'
     show
         ConnectorConfig,
@@ -26,7 +24,7 @@ export 'src/common/common_library.dart'
 export 'src/core/empty_serializer.dart' show emptySerializer;
 export 'src/core/ref.dart'
     show MutationRef, OperationRef, OperationResult, QueryRef, QueryResult;
-
+export 'src/firebase_data_connect.dart';
 export 'src/optional.dart'
     show
         Optional,
@@ -36,4 +34,3 @@ export 'src/optional.dart'
         listDeserializer,
         listSerializer;
 export 'src/timestamp.dart' show Timestamp;
-export 'src/any_value.dart' show AnyValue, defaultSerializer;