+++ packages/firebase_data_connect/firebase_data_connect/lib/src/core/ref.dart
@@ -35,8 +35,14 @@ class QueryResult<Data, Variables> extends OperationResult<Data, Variables> {
 /// Contains variables, transport to execute queries, and serialization/deserialization strategies.
 abstract class OperationRef<Data, Variables> {
   /// Constructor
-  OperationRef(this.dataConnect, this.operationName, this._transport,
-      this.deserializer, this.serializer, this.variables);
+  OperationRef(
+    this.dataConnect,
+    this.operationName,
+    this._transport,
+    this.deserializer,
+    this.serializer,
+    this.variables,
+  );
   Variables? variables;
   String operationName;
   DataConnectTransport _transport;
@@ -48,8 +54,13 @@ abstract class OperationRef<Data, Variables> {
 
   Future<OperationResult<Data, Variables>> execute();
   Future<bool> _shouldRetry() async {
-    String? newToken =
-        await this.dataConnect.auth?.currentUser?.getIdToken(false);
+    String? newToken;
+    try {
+      newToken = await this.dataConnect.auth?.currentUser?.getIdToken();
+    } catch (e) {
+      // Don't retry if there was an issue getting the ID Token.
+      log('There was an error attempting to retrieve the ID Token: $e');
+    }
     bool shouldRetry = newToken != null && _lastToken != newToken;
     _lastToken = newToken;
     return shouldRetry;
@@ -66,14 +77,20 @@ class QueryManager {
   /// Keeps track of what queries are currently active.
   Map<String, Map<String, StreamController<dynamic>>> trackedQueries = {};
   bool containsQuery<Variables>(
-      String queryName, Variables variables, String varsAsStr) {
+    String queryName,
+    Variables variables,
+    String varsAsStr,
+  ) {
     String key = varsAsStr;
     return (trackedQueries[queryName] != null) &&
         trackedQueries[queryName]![key] != null;
   }
 
   Stream addQuery<Variables>(
-      String queryName, Variables variables, String varsAsStr) {
+    String queryName,
+    Variables variables,
+    String varsAsStr,
+  ) {
     // TODO(mtewani): Replace with more stable encoder
     String key = varsAsStr;
     if (trackedQueries[queryName] == null) {
@@ -86,38 +103,48 @@ class QueryManager {
   }
 
   Future<void> triggerCallback<Data, Variables>(
-      String operationName,
-      String varsAsStr,
-      QueryRef<Data, Variables> ref,
-      Data? data,
-      Exception? error) async {
+    String operationName,
+    String varsAsStr,
+    QueryRef<Data, Variables> ref,
+    Data? data,
+    Exception? error,
+  ) async {
     String key = varsAsStr;
     if (trackedQueries[operationName] == null ||
         trackedQueries[operationName]![key] == null) {
       return;
     }
+    // ignore: close_sinks
     StreamController stream = trackedQueries[operationName]![key]!;
-    // TODO(mtewani): Prevent this from getting called multiple times.
-    stream.onCancel = () => stream.close();
-    if (error != null) {
-      stream.addError(error);
-    } else {
-      stream.add(QueryResult<Data, Variables>(dataConnect, data as Data, ref));
+
+    if (!stream.isClosed) {
+      if (error != null) {
+        stream.addError(error);
+      } else {
+        stream
+            .add(QueryResult<Data, Variables>(dataConnect, data as Data, ref));
+      }
     }
   }
 }
 
 class QueryRef<Data, Variables> extends OperationRef<Data, Variables> {
   QueryRef(
-      FirebaseDataConnect dataConnect,
-      String operationName,
-      DataConnectTransport transport,
-      Deserializer<Data> deserializer,
-      this._queryManager,
-      Serializer<Variables> serializer,
-      Variables? variables)
-      : super(dataConnect, operationName, transport, deserializer, serializer,
-            variables);
+    FirebaseDataConnect dataConnect,
+    String operationName,
+    DataConnectTransport transport,
+    Deserializer<Data> deserializer,
+    this._queryManager,
+    Serializer<Variables> serializer,
+    Variables? variables,
+  ) : super(
+          dataConnect,
+          operationName,
+          transport,
+          deserializer,
+          serializer,
+          variables,
+        );
 
   QueryManager _queryManager;
 
@@ -132,24 +159,37 @@ class QueryRef<Data, Variables> extends OperationRef<Data, Variables> {
           e.code == DataConnectErrorCode.unauthorized.toString()) {
         return this.execute();
       } else {
-        throw e;
+        rethrow;
       }
-    } catch (e) {
-      throw e;
     }
   }
 
   Future<QueryResult<Data, Variables>> _executeOperation(String? token) async {
     try {
       Data data = await _transport.invokeQuery<Data, Variables>(
-          operationName, deserializer, serializer, variables, token);
+        operationName,
+        deserializer,
+        serializer,
+        variables,
+        token,
+      );
       QueryResult<Data, Variables> res = QueryResult(dataConnect, data, this);
-      await _queryManager.triggerCallback<Data, Variables>(operationName,
-          serializer(variables as Variables), this, res.data, null);
+      await _queryManager.triggerCallback<Data, Variables>(
+        operationName,
+        serializer(variables as Variables),
+        this,
+        res.data,
+        null,
+      );
       return res;
     } on Exception catch (e) {
       await _queryManager.triggerCallback<Data, Variables>(
-          operationName, serializer(variables as Variables), this, null, e);
+        operationName,
+        serializer(variables as Variables),
+        this,
+        null,
+        e,
+      );
       rethrow;
     }
   }
@@ -164,7 +204,7 @@ class QueryRef<Data, Variables> extends OperationRef<Data, Variables> {
         this.execute();
       } catch (_) {
         // Call to `execute` should properly pass the error to the Stream.
-        log("Error thrown by execute. The error will propagate via onError.");
+        log('Error thrown by execute. The error will propagate via onError.');
       }
     }
     return res;
@@ -179,8 +219,14 @@ class MutationRef<Data, Variables> extends OperationRef<Data, Variables> {
     Deserializer<Data> deserializer,
     Serializer<Variables> serializer,
     Variables? variables,
-  ) : super(dataConnect, operationName, transport, deserializer, serializer,
-            variables);
+  ) : super(
+          dataConnect,
+          operationName,
+          transport,
+          deserializer,
+          serializer,
+          variables,
+        );
 
   @override
   Future<OperationResult<Data, Variables>> execute() async {
@@ -196,17 +242,21 @@ class MutationRef<Data, Variables> extends OperationRef<Data, Variables> {
           e.code == DataConnectErrorCode.unauthorized.toString()) {
         return this.execute();
       } else {
-        throw e;
+        rethrow;
       }
-    } catch (e) {
-      throw e;
     }
   }
 
   Future<OperationResult<Data, Variables>> _executeOperation(
-      String? token) async {
+    String? token,
+  ) async {
     Data data = await _transport.invokeMutation<Data, Variables>(
-        operationName, deserializer, serializer, variables, token);
+      operationName,
+      deserializer,
+      serializer,
+      variables,
+      token,
+    );
     return OperationResult(dataConnect, data, this);
   }
 }