+++ packages/firebase_analytics/firebase_analytics_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.5.10+5
+
+ - Update a dependency to the latest release.
+
+## 0.5.10+4
+
+ - Update a dependency to the latest release.
+
+## 0.5.10+3
+
+ - Update a dependency to the latest release.
+
 ## 0.5.10+2
 
  - Update a dependency to the latest release.