+++ packages/firebase_app_installations/firebase_app_installations_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.6+5
+
+ - Update a dependency to the latest release.
+
+## 0.1.6+4
+
+ - Update a dependency to the latest release.
+
+## 0.1.6+3
+
+ - Update a dependency to the latest release.
+
 ## 0.1.6+2
 
  - Update a dependency to the latest release.