+++ packages/firebase_auth/firebase_auth_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 5.13.5
+
+ - Update a dependency to the latest release.
+
+## 5.13.4
+
+ - Update a dependency to the latest release.
+
+## 5.13.3
+
+ - Update a dependency to the latest release.
+
 ## 5.13.2
 
  - Update a dependency to the latest release.