+++ packages/firebase_messaging/firebase_messaging_web/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 3.9.5
+
+ - Update a dependency to the latest release.
+
+## 3.9.4
+
+ - Update a dependency to the latest release.
+
+## 3.9.3
+
+ - Update a dependency to the latest release.
+
 ## 3.9.2
 
  - Update a dependency to the latest release.